1)Deleted JWT.js in directory config; 2)Added more full method getSignedJwtToken in model of user(in file User.js); 3) Refactored user.js
